requirements.txt
Pillow==8.1.0
pyglet==1.5.15
